moving common funcs used by all policy types to a new file
