+++ pkg/netpol/eval/internal/k8s/netpol.go
@@ -59,92 +59,11 @@ type NetworkPolicy struct {
 // @todo need a network policy collection type along with convenience methods?
 // 	if so, also consider concurrent access (or declare not goroutine safe?)
 
-const (
-	portBase    = 10
-	portBits    = 32
-	egressName  = "Egress"
-	ingressName = "Ingress"
-)
-
-func getProtocolStr(p *v1.Protocol) string {
-	if p == nil || string(*p) == "" { // If not specified, this field defaults to TCP.
-		return string(v1.ProtocolTCP)
-	}
-	return string(*p)
-}
-
-// getPortsRange given a rule port and dest peer, returns: the start and end port numbers,
-// or the port name if it is a named port
-// if input port is a named port, and the dst peer is nil or does not have a matching named port defined, returns
-// an empty range represented by (-1,-1) with the named port string
-func getPortsRange(port *intstr.IntOrString, protocol *v1.Protocol, endPort *int32, dst Peer, policyName string) (start, end int64,
-	portName string, err error) {
-	if port.Type == intstr.String { // rule.Port is namedPort
-		if endPort != nil { // endPort field may not be defined with named port
-			return start, end, "", netpolErr(policyName, alerts.NamedPortErrTitle, alerts.EndPortWithNamedPortErrStr)
-		}
-		ruleProtocol := getProtocolStr(protocol)
-		portName = port.StrVal
-		if dst == nil {
-			// dst is nil, so the namedPort can not be converted, return string port name
-			return common.NoPort, common.NoPort, portName, nil
-		}
-		if dst.PeerType() != PodType {
-			// namedPort is not defined for IP-blocks
-			return start, end, "", netpolErr(policyName, alerts.NamedPortErrTitle, alerts.ConvertNamedPortErrStr)
-		}
-		podProtocol, podPortNum := dst.GetPeerPod().ConvertPodNamedPort(portName)
-		if podProtocol == "" && podPortNum == common.NoPort {
-			// there is no match for the namedPort in the configuration of the pod, return the portName string as "ports range"
-			return common.NoPort, common.NoPort, portName, nil
-		}
-		if podProtocol != ruleProtocol {
-			// the pod has a matching namedPort, but not on same protocol as the rule's protocol; so it can not be converted,
-			// return the string named-port as the "ports range"
-			return common.NoPort, common.NoPort, portName, nil
-		}
-		// else, found match for the rule's named-port in the pod's ports, so it may be converted to port number
-		start = int64(podPortNum)
-		end = int64(podPortNum)
-	} else { // rule.Port is number
-		start = int64(port.IntVal)
-		end = start
-		if endPort != nil {
-			end = int64(*endPort)
-		}
-	}
-	return start, end, portName, nil
-}
-
-func isEmptyPortRange(start, end int64) bool {
-	// an empty range when:
-	// - end is smaller than start
-	// - end or start is not in the legal range (a legal port is 1-65535)
-	return (start < common.MinPort || end < common.MinPort) ||
-		(end < start) ||
-		(start > common.MaxPort || end > common.MaxPort)
-}
-
 func (np *NetworkPolicy) rulePeersAndPorts(ruleIdx int, isIngress bool) ([]netv1.NetworkPolicyPeer, []netv1.NetworkPolicyPort) {
 	if isIngress {
 		return np.Spec.Ingress[ruleIdx].From, np.Spec.Ingress[ruleIdx].Ports
 	}
 	return np.Spec.Egress[ruleIdx].To, np.Spec.Egress[ruleIdx].Ports
-}
-
-// doesRulePortContain gets protocol and port numbers of a rule and other protocol and port;
-// returns if other is contained in the rule's port
-func doesRulePortContain(ruleProtocol, otherProtocol string, ruleStartPort, ruleEndPort, otherPort int64) bool {
-	if !strings.EqualFold(ruleProtocol, otherProtocol) {
-		return false
-	}
-	if isEmptyPortRange(ruleStartPort, ruleEndPort) {
-		return false
-	}
-	if otherPort >= ruleStartPort && otherPort <= ruleEndPort {
-		return true
-	}
-	return false
 }
 
 func (np *NetworkPolicy) ruleConnections(rulePorts []netv1.NetworkPolicyPort, dst Peer,
@@ -217,14 +136,6 @@ func (np *NetworkPolicy) ruleConnections(rulePorts []netv1.NetworkPolicyPort, ds
 			capturedPeersButUnmatchedNamedPortExpl(np.FullName(), ruleName, policyPeerStr, rulePeerStr), isIngress)
 	}
 	return res, nil
-}
-
-// isPeerRepresentative  determines if the peer's source is representativePeer; i.e. its pod fake and has RepresentativePodName
-func isPeerRepresentative(peer Peer) bool {
-	if peer.GetPeerPod() == nil {
-		return false
-	}
-	return peer.GetPeerPod().IsPodRepresentative()
 }
 
 func (np *NetworkPolicy) saveNetpolWarning(warning string) {
@@ -309,31 +220,6 @@ func (np *NetworkPolicy) ruleSelectsPeer(rulePeers []netv1.NetworkPolicyPeer, pe
 	return false, nil
 }
 
-func checkSelectorsMatchForPeer(nsSelector, podSelector *metav1.LabelSelector, peer Peer, policyNs string) (match bool, err error) {
-	if peer.PeerType() == IPBlockType {
-		return false, nil // assuming that peer of type IP cannot be selected by ns and pod selector
-	}
-	// peer is a pod
-	nsSelMatch := false
-	if nsSelector == nil {
-		nsSelMatch = (policyNs == peer.GetPeerPod().Namespace)
-	} else { // namespaceSelector is not nil
-		nsSelMatch, err = doesNamespaceSelectorMatchesPeer(nsSelector, peer)
-		if err != nil {
-			return false, err
-		}
-	}
-	if !nsSelMatch { // namespace selector does not match - no need to check podSelector too
-		return false, nil
-	}
-	// getting here means nsSelMatch is true - lets check podSelector's match for the peer
-	if podSelector == nil {
-		return true, nil
-	}
-	return selectorsMatch(podSelector, peer.GetPeerPod().RepresentativePodLabelSelector,
-		peer.GetPeerPod().Labels, isPeerRepresentative(peer))
-}
-
 // IngressAllowedConn returns true  if the given connections from src to any of the pods captured by the policy is allowed
 func (np *NetworkPolicy) IngressAllowedConn(src Peer, protocol, port string, dst Peer) (bool, error) {
 	// iterate list of rules: []NetworkPolicyIngressRule
@@ -381,46 +267,6 @@ func (np *NetworkPolicy) EgressAllowedConn(dst Peer, protocol, port string) (boo
 		}
 	}
 	return false, nil
-}
-
-const (
-	capturedButNotSelectedExpl   = "selects %s, but %s is not allowed by any %s rule"
-	noMatchExplFormat            = "%s selects %s, and %s selects %s, %s"
-	noXgressRulesExpl            = capturedButNotSelectedExpl + " (no rules defined)"
-	explNoMatchOfNamedPortsToDst = "but named ports of the rule have no match in the configuration of the destination peer"
-)
-
-// ConstPeerString returns pod's owner-name not the peer instance name unless it is ip-block (used for explanation)
-func ConstPeerString(peer Peer) string {
-	peerStr := peer.String()
-	if peer.PeerType() != IPBlockType {
-		peerStr = (&WorkloadPeer{peer.GetPeerPod()}).String()
-	}
-	return peerStr
-}
-
-func directionName(isIngress bool) string {
-	if isIngress {
-		return ingressName
-	}
-	return egressName
-}
-
-func notSelectedByRuleExpl(isIngress bool, policyName, expl, policyPeerStr, rulePeerStr string) string {
-	return fmt.Sprintf("%s "+expl, policyName, policyPeerStr, rulePeerStr, directionName(isIngress))
-}
-
-func capturedPeersButUnmatchedConnsExpl(policyName, ruleName, policyPeerStr, rulePeerStr string) string {
-	return fmt.Sprintf(noMatchExplFormat, policyName, policyPeerStr, ruleName, rulePeerStr,
-		common.ExplNotReferencedProtocolsOrPorts)
-}
-
-func capturedPeersButUnmatchedNamedPortExpl(policyName, ruleName, policyPeerStr, rulePeerStr string) string {
-	return fmt.Sprintf(noMatchExplFormat, policyName, policyPeerStr, ruleName, rulePeerStr, explNoMatchOfNamedPortsToDst)
-}
-
-func allowedByRuleExpl(policyName, ruleName string) string {
-	return fmt.Sprintf("%s allows connections by %s", policyName, ruleName)
 }
 
 // GetXgressAllowedConns returns the set of allowed connections to a captured dst pod from the src peer (for Ingress)
@@ -475,10 +321,6 @@ func (np *NetworkPolicy) GetXgressAllowedConns(src, dst Peer, isIngress bool) (*
 
 func (np *NetworkPolicy) netpolWarning(description string) string {
 	return fmt.Sprintf("%s: %s", np.FullName(), description)
-}
-
-func netpolErr(policyName, title, description string) error {
-	return fmt.Errorf("%s %s: %s", policyName, title, description)
 }
 
 func (np *NetworkPolicy) parseNetpolCIDR(cidr string, except []string) (*netset.IPBlock, error) {
@@ -581,10 +423,6 @@ func (np *NetworkPolicy) Selects(p *Pod, direction netv1.PolicyType) (bool, erro
 
 func (np *NetworkPolicy) FullName() string {
 	return fmt.Sprintf("NetworkPolicy '%s'", types.NamespacedName{Name: np.Name, Namespace: np.Namespace}.String())
-}
-
-func ruleName(ruleIdx int, isIngress bool) string {
-	return fmt.Sprintf("%s rule #%d", directionName(isIngress), ruleIdx+1)
 }
 
 func (np *NetworkPolicy) LogWarnings(l logger.Logger) []string {
